Refactor electron range module for backward compatibility and add comprehensive tests
